Add files via upload
ss.png - Sample of the screenshot taken with Selenium that was processed and cropped
data.csv - Program output
driver.py - Code for the program - I might separate the code and redo the organization, but for the foreseeable future this will suit my needs, as all I may need is a .csv file for the next project
chromedriver.exe - This project won't work without the right version. (download corresponding version of ChromeDriver for Google Chrome that you have downloaded)
Note: Line 12 is the PATH for Pytesseract, you will need to do "pip install pytesseract" in order to use this program. You may need to update the path on Line 12, but most likely not.